python_tools/text_replace/requirements.txt
paddlepaddle
paddleocr
opencv-python
pillow
numpy
gradio
torch
diffusers
transformers
accelerate
segment-anything
